backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq
openai-whisper
gtts==2.4.0
pydub==0.25.1
opencv-python
librosa==0.10.1
tensorflow
numpy
python-multipart==0.0.6
websockets==12.0
